+++ src/pyramid_bimt/static/bimt.js
@@ -49,8 +49,6 @@
             });
         });
 
-        // Activate chosen jquery plugin for all selects
-        $("select").chosen({search_contains: true});
     }
 
     $(document).ready(function () {